Readd graphQLFetch and useBlockPreviewFetch to site-react
