+++ packages/site/site-nextjs/src/graphQLFetch/graphQLFetch.ts
@@ -1,56 +1,8 @@
+import { convertPreviewDataToHeaders } from "@comet/site-react";
+
 import { type SitePreviewData } from "../sitePreview/SitePreviewUtils";
 
 type Fetch = typeof fetch;
-
-export function convertPreviewDataToHeaders(previewData?: SitePreviewData) {
-    const { includeInvisibleBlocks, includeInvisiblePages } = {
-        includeInvisiblePages: !!previewData,
-        includeInvisibleBlocks: previewData && previewData.includeInvisible,
-    };
-
-    const headers: Record<string, string> = {};
-    const includeInvisibleContentHeaderEntries: string[] = [];
-    if (includeInvisiblePages) {
-        includeInvisibleContentHeaderEntries.push("Pages:Unpublished");
-    }
-    if (includeInvisibleBlocks) {
-        includeInvisibleContentHeaderEntries.push("Blocks:Invisible");
-    }
-    // tells api to send invisble content
-    // authentication is required when this header is used
-    if (includeInvisibleContentHeaderEntries.length > 0) {
-        headers["x-include-invisible-content"] = includeInvisibleContentHeaderEntries.join(",");
-    }
-    return headers;
-}
-
-// Adapted from https://github.com/jasonkuhrt/graffle/blob/main/src/layers/6_helpers/gql.ts
-export const gql = (chunks: TemplateStringsArray, ...variables: unknown[]): string => {
-    return chunks.reduce((acc, chunk, index) => {
-        let variable;
-
-        if (index in variables) {
-            if (typeof variables[index] !== "string") {
-                let errorMessage =
-                    `Non-string variable in the GraphQL document\n\n` +
-                    `This is most likely due to importing a GraphQL document from a React Client Component.\n` +
-                    `All GraphQL documents need to be imported from React Server Components (i.e. no "use client" notation).`;
-
-                if (chunk.trim().length > 0) {
-                    errorMessage += `\n\nThe error occurred in the following GraphQL document:\n${chunk}`;
-                }
-
-                throw new Error(errorMessage);
-            } else {
-                variable = variables[index];
-            }
-        } else {
-            variable = "";
-        }
-
-        return `${acc}${chunk}${variable}`;
-    }, ``);
-};
 
 export function createFetchWithPreviewHeaders(fetch: Fetch, previewData?: SitePreviewData): Fetch {
     const defaultHeaders = convertPreviewDataToHeaders(previewData);
@@ -71,55 +23,5 @@ export function createFetchWithDefaults(fetch: Fetch, defaults: RequestInit): Fe
                 ...init?.next,
             },
         });
-    };
-}
-
-export type GraphQLFetch = <T, V>(query: string, variables?: V, init?: RequestInit) => Promise<T>;
-
-export function createGraphQLFetch(fetch: Fetch, url: string): GraphQLFetch {
-    return async function <T, V>(query: string, variables?: V, init?: RequestInit): Promise<T> {
-        let response;
-        if (init?.method === "GET") {
-            const fetchUrl = new URL(url);
-            fetchUrl.searchParams.append("query", query);
-            fetchUrl.searchParams.append("variables", JSON.stringify(variables));
-            response = await fetch(fetchUrl, init);
-        } else {
-            response = await fetch(url, {
-                method: "POST",
-                ...init,
-                headers: { "Content-Type": "application/json", ...init?.headers },
-                body: JSON.stringify({
-                    query,
-                    variables,
-                }),
-            });
-        }
-        if (!response.ok) {
-            let errorMessage = `Network response was not ok. Status: ${response.status}`;
-            const body = await response.text();
-
-            try {
-                const json = JSON.parse(body);
-
-                const { errors } = json;
-                if (errors) {
-                    errorMessage += `\n\nGraphQL error(s):\n- ${errors.map((error: { message: string }) => error.message).join("\n- ")}`;
-                }
-            } catch (error) {
-                errorMessage += `\n${body}`;
-            }
-
-            throw new Error(errorMessage);
-        }
-
-        const { data, errors } = await response.json();
-
-        if (errors) {
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-            throw new Error(`GraphQL error(s):\n- ${errors.map((error: any) => error.message).join("\n- ")}`);
-        }
-
-        return data;
     };
 }
